Fixed linters and some problems
