+++ index.css
@@ -1,193 +1,193 @@
 * {
-    box-sizing: border-box;
-    font-weight: 600;
-    text-decoration: none;
-    list-style: none;
-    color: #000;
+  box-sizing: border-box;
+  font-weight: 600;
+  text-decoration: none;
+  list-style: none;
+  color: #000;
 }
 
 body {
-    margin: 0;
-    padding: 0;
+  margin: 0;
+  padding: 0;
 }
 
 .nav {
-    width: 100%;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    padding: 0 1.2em;
-    border: 3px solid black;
+  width: 100%;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 0 1.2em;
+  border: 3px solid black;
 }
 
 .nav h4 {
-    animation-name: flip;
-    animation-duration: 2s;
-    animation-timing-function: ease-in-out;
-    animation-iteration-count: infinite;
-    animation-play-state: running;
-    font-size: 1.4em;
-    color: brown;
+  animation-name: flip;
+  animation-duration: 2s;
+  animation-timing-function: ease-in-out;
+  animation-iteration-count: infinite;
+  animation-play-state: running;
+  font-size: 1.4em;
+  color: brown;
 }
 
 header {
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-    align-items: end;
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: end;
 }
 
 time {
-    padding: 5px 20px;
+  padding: 5px 20px;
 }
 
 @keyframes flip {
-    0% {
-        opacity: 0.2;
-        transform: scale(-1.1);
-    }
+  0% {
+    opacity: 0.2;
+    transform: scale(-1.1);
+  }
 
-    100% {
-        opacity: 1;
-        font-weight: 800;
-        transform: scale(1.2);
-        color: rgb(255, 60, 0);
-    }
+  100% {
+    opacity: 1;
+    font-weight: 800;
+    transform: scale(1.2);
+    color: rgb(255, 60, 0);
+  }
 }
 
 #nav {
-    display: flex;
-    gap: 0.8em;
-    font-size: 1.2em;
+  display: flex;
+  gap: 0.8em;
+  font-size: 1.2em;
 }
 
 a {
-    transition: all 0.6s ease-out;
-    color: brown;
+  transition: all 0.6s ease-out;
+  color: brown;
 }
 
 a:hover {
-    font-size: 1.25em;
-    color: rgb(22, 186, 184);
+  font-size: 1.25em;
+  color: rgb(22, 186, 184);
 }
 
 #bookList li {
-    margin: 4px 0;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    text-align: center;
+  margin: 4px 0;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  text-align: center;
 }
 
 #nav li:not(:last-child) {
-    border-right: 3px solid black;
-    padding: 0 0.4em;
+  border-right: 3px solid black;
+  padding: 0 0.4em;
 }
 
 .parent {
-    width: 100%;
-    min-height: 100%;
-    display: flex;
-    flex-direction: column;
-    justify-content: start;
-    align-items: center;
+  width: 100%;
+  min-height: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: start;
+  align-items: center;
 }
 
 h1,
 h3 {
-    font-size: 2.6rem;
+  font-size: 2.6rem;
 }
 
 h6 {
-    font-size: 1.3rem;
-    color: chocolate;
+  font-size: 1.3rem;
+  color: chocolate;
 }
 
 #bookList {
-    width: 100%;
-    list-style: none;
-    padding: 0 10px;
-    display: flex;
-    flex-direction: column;
-    border: 4px solid black;
+  width: 100%;
+  list-style: none;
+  padding: 0 10px;
+  display: flex;
+  flex-direction: column;
+  border: 4px solid black;
 }
 
 .bookslist {
-    display: none;
-    flex-direction: column;
-    min-width: 70%;
-    align-items: center;
-    margin-top: 2em;
+  display: none;
+  flex-direction: column;
+  min-width: 70%;
+  align-items: center;
+  margin-top: 2em;
 }
 
 h2 {
-    display: flex;
-    justify-content: space-between;
-    margin-left: 5px;
+  display: flex;
+  justify-content: space-between;
+  margin-left: 5px;
 }
 
 hr {
-    width: 200px;
-    border: 1px solid;
+  width: 200px;
+  border: 1px solid;
 }
 
 #titleInput,
 #authorInput {
-    max-width: 24em;
-    margin-bottom: 1em;
-    padding: 0.4em 0;
-    border-radius: 4px;
-    border: 4px solid #000;
+  max-width: 24em;
+  margin-bottom: 1em;
+  padding: 0.4em 0;
+  border-radius: 4px;
+  border: 4px solid #000;
 }
 
 .add {
-    width: fit-content;
-    padding: 0.4em 1.2em;
-    text-align: center;
-    border-radius: 4px;
-    border: 1px solid gray;
-    box-shadow: 4px 4px rgba(0, 0, 0, 1);
-    margin-left: 230px;
-    cursor: pointer;
+  width: fit-content;
+  padding: 0.4em 1.2em;
+  text-align: center;
+  border-radius: 4px;
+  border: 1px solid gray;
+  box-shadow: 4px 4px rgba(0, 0, 0, 1);
+  margin-left: 230px;
+  cursor: pointer;
 }
 
 .removeButtonStyle {
-    display: block;
-    padding: 0.5em 1em;
-    border-radius: 5px;
-    border: 1px solid gray;
-    margin: 0 6px 4px;
-    box-shadow: 4px 4px rgba(0, 0, 0, 1);
-    cursor: pointer;
+  display: block;
+  padding: 0.5em 1em;
+  border-radius: 5px;
+  border: 1px solid gray;
+  margin: 0 6px 4px;
+  box-shadow: 4px 4px rgba(0, 0, 0, 1);
+  cursor: pointer;
 }
 
 .main-page {
-    display: flex;
-    flex-direction: column;
-    margin-top: 2em;
+  display: flex;
+  flex-direction: column;
+  margin-top: 2em;
 }
 
 /* intercalate and change the style of the half of the books tittles  */
 
 .makeMeGrey {
-    background-color: lightgrey;
+  background-color: lightgrey;
 }
 
 #contact {
-    display: none;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    font-size: 1.3rem;
+  display: none;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  font-size: 1.3rem;
 }
 
 .copy-right {
-    border: 3px solid black;
-    font-size: 1.4rem;
-    padding: 0.4em 0;
-    position: fixed;
-    top: 90%;
-    left: 0;
-    right: 0;
-    text-align: center;
+  border: 3px solid black;
+  font-size: 1.4rem;
+  padding: 0.4em 0;
+  position: fixed;
+  top: 90%;
+  left: 0;
+  right: 0;
+  text-align: center;
 }
