+++ index.html
@@ -16,7 +16,7 @@
                 <li><a href="#contact" id="nav-contact">Contact</a></li>
             </ul>
         </nav>
-        <time datetime="2023-06-0T09:30">June 1st 2023, 09:30 am</time>
+        <time id="time"></time>
   </header>
     <main class="parent">
         <div class="bookslist">
